pyproject.toml
[project]
name = "api-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "tzdata>=2025.2",
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery>=3.35.1",
    "google-cloud-bigquery-storage>=2.32.0",
    "colorlog>=6.9.0",
    "colorama>=0.4.6",
    "apscheduler>=3.11.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pylint>=3.3.7",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "ruff==0.1.5",
    "types-colorama>=0.4.15.20240311",
    "types-requests>=2.32.4.20250611",
    "vulture>=2.14",
]

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "venv",
    ".venv",
    ".env",
    ".idea",
    ".vscode",
]
line-length = 79
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    # pycodestyle
    #"E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    #"SIM",
    # isort
    "I",
]
ignore = ["UP006", "UP035", "UP017"]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401"]


[tool.mypy]
strict = false
exclude = [
    "^venv/.*$",
    "^migrations/.*$",
    "^tests/.*$",
    "^scripts/.*$",
]
plugins = [
  "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = "apscheduler.*,pydantic_settings.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
python_files = "test_*.py"
pythonpath = ["."]
#asyncio_mode = "auto"
#asyncio_default_fixture_loop_scope = "function"
#log_cli = true
#log_cli_level = "DEBUG"

[tool.pylint.messages_control]
disable = [
  "C0114",
  "C0115",
  "C0116",
  "R0903",
  "R0913",
  "R0914",
  "R0917",
]
[tool.pylint.master]
ignore = [
   ".venv",
   "migrations",
   "tests",
   "scripts",
]
